requirements.txt
elasticsearch
requests
python-dotenv
flask